fix: keep state of record (draft or published)

+++ repository_cli/cli/records.py
@@ -12,16 +12,30 @@ import json
 import click
 from flask.cli import with_appcontext
 from flask_principal import Identity
-from invenio_rdm_records.proxies import current_rdm_records
+from invenio_rdm_records.records.models import RDMRecordMetadata
 
 from .click_options import option_identifier, option_pid
-from .util import get_identity
+from .util import get_draft, get_identity, get_records_service, update_record
 
 
 @click.group()
 def records():
     """Management commands for records."""
     pass
+
+
+@records.command("list")
+@with_appcontext
+def list_records():
+    """List record's.
+
+    example call:
+        invenio repository records list
+    """
+    records = RDMRecordMetadata.query
+    for index, metadata in enumerate(records):
+        fg = "blue" if index % 2 == 0 else "cyan"
+        click.secho(json.dumps(metadata.data, indent=2), fg=fg)
 
 
 @click.group()
@@ -43,7 +57,7 @@ def list_identifiers(pid):
         invenio repository records identifiers list
     """
     identity = get_identity()
-    service = current_rdm_records.records_service
+    service = get_records_service()
     record_data = service.read(id_=pid, identity=identity).data.copy()
     current_identifiers = record_data["metadata"].get("identifiers", [])
 
@@ -74,14 +88,22 @@ def add_identifier(identifier, pid):
         return
 
     identity = get_identity("system_process")
-    service = current_rdm_records.records_service
+    service = get_records_service()
 
     # get current draft or create new one
-    record_data = service.edit(id_=pid, identity=identity).data.copy()
+    draft = get_draft(pid, identity)
+    should_publish = False
+    if draft is None:
+        should_publish = True
+        draft = service.edit(id_=pid, identity=identity)
+
+    record_data = draft.data.copy()
     current_identifiers = record_data["metadata"].get("identifiers", [])
     current_schemes = [_["scheme"] for _ in current_identifiers]
     scheme = identifier["scheme"]
     if scheme in current_schemes:
+        if should_publish:
+            service.delete_draft(id_=pid, identity=identity)
         click.secho(f"scheme '{scheme}' already in identifiers", fg="red")
         return
 
@@ -89,11 +111,15 @@ def add_identifier(identifier, pid):
     record_data["metadata"]["identifiers"] = current_identifiers
 
     try:
-        service.update_draft(id_=pid, identity=identity, data=record_data)
-        service.publish(id_=pid, identity=identity)
+        update_record(
+            pid=pid,
+            identity=identity,
+            should_publish=should_publish,
+            new_data=record_data,
+            old_data=draft.data,
+        )
         click.secho(pid, fg="green")
     except Exception as e:
-        service.delete_draft(id_=pid, identity=identity)
         click.secho(f"{pid}, {e}", fg="red")
 
     return
@@ -116,10 +142,16 @@ def replace_identifier(identifier, pid):
         return
 
     identity = get_identity("system_process")
-    service = current_rdm_records.records_service
+    service = get_records_service()
 
     # get current draft or create new one
-    record_data = service.edit(id_=pid, identity=identity).data.copy()
+    draft = get_draft(pid, identity)
+    should_publish = False
+    if draft is None:
+        should_publish = True
+        draft = service.edit(id_=pid, identity=identity)
+
+    record_data = draft.data.copy()
     current_identifiers = record_data["metadata"].get("identifiers", [])
     scheme = identifier["scheme"]
     replaced = False
@@ -130,17 +162,21 @@ def replace_identifier(identifier, pid):
             break
 
     if not replaced:
+        if should_publish:
+            service.delete_draft(id_=pid, identity=identity)
         click.secho(f"scheme '{scheme}' not in identifiers", fg="red")
         return
 
     record_data["metadata"]["identifiers"] = current_identifiers
 
     try:
-        service.update_draft(id_=pid, identity=identity, data=record_data)
-        service.publish(id_=pid, identity=identity)
+        update_record(
+            pid=pid,
+            identity=identity,
+            should_publish=should_publish,
+            new_data=record_data,
+            old_data=draft.data,
+        )
         click.secho(pid, fg="green")
     except Exception as e:
-        service.delete_draft(id_=pid, identity=identity)
         click.secho(f"{pid}, {e}", fg="red")
-
-    return
